subsys: shell: Call bypass callback even when no data has been received
The Shell subsystem allows a command to bypass the shell and receive
data directly from the interface on which the shell is spawned. However,
the bypass callback is only called when data is available on the
interface. With this patch, the bypass callback is called even when no
data is received on the interface. For example, when the interface times
out upon reading data. Hence, the polling interval of the bypass
callback is not necessarily deterministic.