drivers: sensor: nordic: qdec_nrfx: patch QDEC instantiation
The current Nordic QDEC driver implementation does not allow
the sampling and reporting period to be configured through devicetree.
It has configured the report period to a fixed 40 samples / report and
the sampling period to 2048 us. The latter is much too slow for our
application: encoder produces 3 PPR @ 20000 RPM = 60000 PPM or 1000 PPS.
But to measure 1000 PPS we need to sample at least at 2000 SPS to detect
both rising as falling edge. Additionally, we have 2 asynchronous hall
sensors so we'll have to sample at at least 4000 SPS. Hence, sampling
at 128 us yields a sampling rate of 7812.5 SPS, which is the smallest
value that is sufficient for our needs. Since we also want to reduce
the latency on the motor control loop, we configure the reporting
period to the lowest possible value which is 1 sample / report. In
these conditions, and with the modified calculated of the absolute
position which is introduced in this commit, the overflow is unlikely
to occur and hence the static assert can be disabled.